# openapi.yaml
openapi: 3.0.3
info:
  title: Billr API
  description: |
    B2B subscription billing API. Manage accounts, plans, subscriptions, usage metering,
    invoices, and webhooks. All monetary values are integers in the smallest currency unit
    (e.g. cents for USD). Never floats.

    ## Authentication
    All endpoints require a Bearer API key in the `Authorization` header, except
    `GET /plans` and `GET /plans/{planId}` which are public.

    ## Rate Limiting
    Requests are rate-limited per API key. Limits are returned in response headers:
    - `X-RateLimit-Limit` — requests allowed per window
    - `X-RateLimit-Remaining` — requests remaining
    - `X-RateLimit-Reset` — UTC epoch seconds when window resets

    ## Pagination
    List endpoints use cursor-based pagination. Pass `after` (a record ID) and `limit`
    (default 20, max 100). If `next_cursor` is present in the response there are more pages.

    ## Idempotency
    POST endpoints that create resources support an `Idempotency-Key` request header.
    Supply a unique UUID per logical operation; duplicate keys within 24 h return the
    cached response without re-executing the operation.
  version: 1.0.0
  contact:
    name: Billr Support
    email: support@billr.io
    url: https://docs.billr.io
  license:
    name: Proprietary
servers:
  - url: https://api.billr.io/v1
    description: Production
security:
  - BearerAuth: []
...
paths:
...
  /subscriptions/{subId}/usage:
    get:
      tags: [Subscriptions]
      operationId: getSubscriptionUsage
      summary: Get subscription usage summary
      description: Returns aggregated usage totals for the current billing period, grouped by metric.
      parameters:
        - $ref: '#/components/parameters/SubId'
      responses:
        '200':
          description: Usage summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ---------------------------------------------------------------------------
  # Usage Metering
  # ---------------------------------------------------------------------------
  /usage:
    post:
      tags: [Usage]
      operationId: ingestUsageRecord
      summary: Ingest usage event
      description: |
        Records a metered usage event. Use `idempotency_key` (a stable UUID for this event)
        to prevent double-counting — duplicate keys within the same billing period are silently
        deduplicated. `recorded_at` defaults to server time if omitted.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUsageRecordRequest'
            example:
              subscription_id: sub_01HX3Y9ZG2VMBPN5Q8T7RWKD
              metric: api_calls
              quantity: 150
              idempotency_key: "evt_20260419_api_calls_batch_001"
              source: "api-gateway"
      responses:
        '201':
          description: Usage record created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageRecord'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

    get:
      tags: [Usage]
      operationId: listUsageRecords
      summary: List usage records
      parameters:
        - name: subscription_id
          in: query
          required: true
          schema:
            type: string
        - name: metric
          in: query
          schema:
            type: string
          description: Filter to a single metric name
        - name: start
          in: query
          schema:
            type: string
            format: date-time
          description: ISO 8601 range start (inclusive)
        - name: end
          in: query
          schema:
            type: string
            format: date-time
          description: ISO 8601 range end (exclusive)
        - $ref: '#/components/parameters/After'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated usage records
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/UsageRecord'
                  next_cursor:
                    type: string
                    nullable: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /usage/summary:
    get:
      tags: [Usage]
      operationId: getUsageSummary
      summary: Get usage summary
      description: Returns per-metric totals for a subscription. Defaults to the current billing period if no date range is provided.
      parameters:
        - name: subscription_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Aggregated usage summary
          content:
            application/json:
              schema:
                type: object
                required: [metrics]
                properties:
                  metrics:
                    type: object
                    description: Keys are metric names; values contain totals and the period window.
                    additionalProperties:
                      type: object
                      required: [total, period_start, period_end]
                      properties:
                        total:
                          type: number
                          description: Total quantity consumed this period
                        period_start:
                          type: string
                          format: date-time
                        period_end:
                          type: string
                          format: date-time
              example:
                metrics:
                  api_calls:
                    total: 48230
                    period_start: "2026-04-01T00:00:00Z"
                    period_end: "2026-04-30T23:59:59Z"
                  active_subscriptions:
                    total: 312
                    period_start: "2026-04-01T00:00:00Z"
                    period_end: "2026-04-30T23:59:59Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ---------------------------------------------------------------------------
  # Invoices
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API key obtained from POST /auth/keys. Prefix with `Bearer `.
  parameters:
...
    SubId:
      name: subId
      in: path
      required: true
      schema:
        type: string
      description: Subscription ID (prefix `sub_`)
...
    After:
      name: after
      in: query
      schema:
        type: string
      description: Cursor for forward pagination — the `id` of the last item on the previous page
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of results per page
  responses:
    BadRequest:
      description: Invalid request body or query parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: validation_error
            message: "'billing_interval' must be one of: monthly, yearly"
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: unauthorized
            message: "Invalid or missing Bearer token"
    Forbidden:
      description: API key does not have permission for this operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: forbidden
            message: "read_only keys cannot perform this action"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: not_found
            message: "Account acc_01HX3Y9ZG2VMBPN5Q8T7RWKD not found"
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until the rate limit resets
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: rate_limited
            message: "Rate limit exceeded. Retry after 30 seconds."
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: internal_error
            message: "An unexpected error occurred. Our team has been notified."
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
          example: not_found
        message:
          type: string
          description: Human-readable description
          example: "Account not found"

    # -------------------------------------------------------------------------
    # Auth
    # -------------------------------------------------------------------------
...
    UsageSummary:
      type: object
      required: [subscription_id, period_start, period_end, metrics]
      properties:
        subscription_id:
          type: string
        period_start:
          type: string
          format: date-time
        period_end:
          type: string
          format: date-time
        metrics:
          type: object
          additionalProperties:
            type: object
            required: [total]
            properties:
              total:
                type: number
              overage_units:
                type: number
                description: Units above the plan's included quantity
              overage_amount:
                type: integer
                description: Overage charge in smallest currency unit

    # -------------------------------------------------------------------------
    # Usage
    # -------------------------------------------------------------------------
    UsageRecord:
      type: object
      required: [id, subscription_id, metric, quantity, idempotency_key, recorded_at, created_at]
      properties:
        id:
          type: string
          example: usg_01HX3Y9ZG2VMBPN5Q8T7RWKD
        subscription_id:
          type: string
        metric:
          type: string
          description: Metric name as defined in the plan (e.g. `api_calls`, `seats`)
          example: api_calls
        quantity:
          type: number
          description: Consumed quantity for this event
          example: 150
        idempotency_key:
          type: string
          description: Client-supplied deduplication key
          example: evt_20260419_api_calls_batch_001
        source:
          type: string
          nullable: true
          description: Optional label for the originating system
          example: api-gateway
        recorded_at:
          type: string
          format: date-time
          description: When the usage actually occurred (may differ from `created_at`)
        created_at:
          type: string
          format: date-time
    CreateUsageRecordRequest:
      type: object
      required: [subscription_id, metric, quantity, idempotency_key]
      properties:
        subscription_id:
          type: string
        metric:
          type: string
        quantity:
          type: number
          minimum: 0
        idempotency_key:
          type: string
          maxLength: 255
        source:
          type: string
          maxLength: 100
        recorded_at:
          type: string
          format: date-time

    # -------------------------------------------------------------------------
    # Invoices
    # -------------------------------------------------------------------------
...